+++ test/model_test.rb
@@ -252,7 +252,12 @@ class ModelTest < Minitest::Test
 
   def test_inspect
     user = User.create!(email: "test@example.org")
+    assert_includes user.inspect, "email: [FILTERED]"
+    refute_includes user.inspect, "test@example.org"
+
+    user = User.select(:id).last
     refute_includes user.inspect, "email"
+    refute_includes user.inspect, "test@example.org"
   end
 
   def test_serializable_hash